Squibble and on trigger
