+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6b612921badda7b4ca51935e3e9aa72f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
