+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: accef4151132e4f46a22975e10c7bfa8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
