+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Attack.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4705b1ad7131d3c409f1d492cdd3d5c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
