+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Death.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a717f952620182d4895740dec1e49bb0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
