+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Hurt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6db766415c501df4ab39f6c98c35fb3c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
