+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Idle.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4146c4f2eb60fde48ae17e165d954d1f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
