+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Run.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c151ed0201109db489babb2f3f7a7539
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
