+++ the mars delivery/Assets/Omar/Assets/Sprites/Slime3/Walk.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a71aafce19acba4b8fd0d5afa84a879
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
