+++ the mars delivery/Assets/Omar/Scenes/combat.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0b8a31e596c964e48879100ae6d149a9
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
